ejercicios/crewai/pyproject.toml
[project]
name = "crewai-exercises"
version = "0.1.0"
description = "CrewAI exercises for data analysis with multi-agent systems"
requires-python = ">=3.9"
dependencies = [
    "crewai",
    "pandas>=1.5.0",
    "requests>=2.28.0",
    "pillow>=9.0.0",
    "numpy>=1.20.0",
    "langchain-community",
    "python-dotenv",
]
